Updated Caching + Offender Iteration Start
This commit moves caching from the Main execution into the GetURL
function (so it's now smart enough to cache on its own based on URLs).
I used a similar scheme to what CodeIgniter does for this
functionality. (md5 the urls).

This commit also introduces new code which iterates and pulls content
for inmate detail pages.